Add mock -f switch: loads file and defers exec of main
I want to simplify argument handling.  The new -f switch
uses the existing needs word to load a file.  If that file
provides a main word, then it will be used in place of the
interactive loop.  Any mock switches defined in the file
can be called by the remainder of the command line.
The main word is called after argument handling is done.

variable srcfile 80 allot
: -in`  wsparse srcfile append ;
: cat   openr dup>r ?ior START here swap 1 write ?ior ENTER
        here 4096 r read dup ?ior 0- 0= UNTIL drop rdrop ;
: main  srcfile c@+ cat ;

$ ffcat -in /etc/motd

Signed-off-by: Dan Good <dan@dancancode.com>