Permit turnkey to disable normal argument eval
The default behavior is to treat whitespace separated tokens
in argv as Forth words.  This commit factors the argument
processing into a vector, and adds a word to disable a vector.
Since argument processing will have already happened before
a script can disable it, this is intended for turnkey use.

$ cat ex1.ff
doargv n^
: main 2dup type cr ."this is main^J" ;
$ ./ff -f ex1.ff -f mkimage; make fftk; mv fftk ex1
$ ./ex1
./ex1
this is main
$ ./ex1 a b c
./ex1abc
this is main
$ ./ex1 a b c | cat -v
./ex1^@a^@b^@c
this is main

Added the word zlen ( @ -- @ # ) implemented in assembly to
count the characters in a nul terminated string.

Signed-off-by: Dan Good <dan@dancancode.com>