Simplify auto termination
Remove the check for the '\' character. Now the terminating space and
semicolon will be added after the '\', but that doesn't matter since
the \` word discards the rest of the line - so the end result is the
same as before.

The prior version did not normalize the flags in the check and had
potential ambiguity should accept return a 1 byte long input.  This
commit is free of those issues.

Signed-off-by: Dan Good <dan@dancancode.com>