Add stat and mmap related words
There's mmapr for read only access and mmapw for read/write.
To make a new map of a given size, there's mmapwt.  Flags
may be specified in the mm structure supplied on the stack.
?ior is changed so as not to mistake addresses for errors.

Here's yet another turnkey cat, with its own ?ior that skips
to the next loop iteration on an error.

doargv n^
mkmm m
: ?ior  dup $FF | -1 = 2drop IF strerror rdrop ELSE drop THEN ;
: cat   m mmapr ?ior m @ m mm.sz @ type m munmap 2drop ;
: main  0 argc 1- TIMES 1+ dup argv cat REPEAT ;

$ ./ff -f full.ff -f cat3.ff -f mkimage; make fftk; mv fftk cat3

Signed-off-by: Dan Good <dan@dancancode.com>